+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.10.0
+
+-   Added `throughputLimiter` - Limit throughput (items/sec).
+-   Added `sleep` - Sleep for `time` ms before resolving the Promise.
+-   `batchQueue` - Added `maxItemsPerSec` and `maxBytesPerSec` options.
+-   Changed testing library from jest to vitest.
+
 # 0.9.0
 
 -   Better types - removed all use of `any`.